feat: add a `parser` module for yaml parser
This module allows the use of yaml file and cache on it persistently to
avoid huge fs cost. Most base16 users at some point might already have
some setup that utilises single base16 yml file and apply to the whole
system, and here's where this comes in handy without the need to copy
and paste all the color values or manual fs and parse the content in lua
by hand.

Note that this module is totally optional and lazy loaded only when it
is needed.